Convert project to Gradle project for easy management
When this project was first initialized, it was developed using NetBeans and adding paths to classPath manually. Now that I've learned much more and prefer classPath, etc. to be defined in a config file for a simpler clone-and-run existence (compounded by the fact that other IDEs are used now instead), add a Gradle configuration so the project can be run and built from anywhere with any system configuration.